shapes: _Font.bold = None clears bold attribute
The 'b' attribute on rPr is actually a tri-state setting. It can be
"true/1", "false/0", or not present. When not present, the bold setting
is inherited. Adding None as a valid setting adds the option of clearing
the value to the prior options of setting it either True or False.