oxml: extract serialize_part_xml()
A general purpose function, oxml_tostring(), was being used to
serialize the XML for a part before saving. This placed a lot of
unnecessary lxml API choices in the part code (in the form of parameter
choices on oxml_tostring(). Creating a special-purpose function for
this single job greatly simplifies the interface and conforms to the
"do one thing" principle.