util: refactor Pt to use EMU natively
Formerly, points were a non-uniform length value, being held internally
in hundredths of a point. This made a certain amount of sense because
font sizes are stored in Open XML as hundredths of a point. This patch
changes them to be stored internally in English Metric Units, and
conversion to centipoints is performed during XML read/write.