oxml: convert CT_Placeholder.type to xmlchemy
Entailed a lot of changes because the interface to ph.type went from the
ST_PlaceholderType string to the PP_PLACEHOLDER enumeration value. That
interface was fairly widely used.

It turned up the smell that creating a placeholder shape should probably
use a PP_PLACEHOLDER instead of low-level specification of orient, sz,
etc., which are implied by the type. So that's something to clean up if
back in there later for a deep-dive around placeholders.