oxml: convert CT_TextParagraph to xmlchemy
Ended up adding more than I removed, but should be in line for future
replacement with ZeroOrMoreChoice() when that's implemented. Couldn't
implement it here because there's no test support yet for <a:br> and
<a:fld> elements and didn't want to take the time to add that just now,
particularly because there's no API to use that interface; so probably
not the best archetypal example to derive a general solution from.