xmlch: remove None case on OptAttr setter
Original behavior was to always accept None as a setting to remove the
attribute. Often this was the same as _default, because _default
defaults to None, but to prevent unexpected API behavior, only the
specified default value should remove the attribute. In cases where
aliases for that value are desired in the API, they should be mapped to
the registered default in the API method.