prs: rewrite CT_SlideIdList._next_id
The original version filled in gaps in the slide numbering. This
interferes with the slide-reference-conservation behavior of slide ID,
in particular that a new slide could have the same ID as a deleted
slide. By assigning above the maximum assigned slide ID, the chance of
reassigning the id of a deleted slide is minimized (but not eliminated).

Also:
* remove obsolete tests/oxml/unitdata/presentation.py
* remove obsolete tests/oxml/unitdata/slides.py