docs: fix ReadTheDocs problem w/embedded link
A published "fix" to allow external links to appear without a warning
breaks with a recent update to Sphinx. This fix was included in the
conf.py module to allow the TravisCI badge to be displayed on the root
page.

Update that fix as described here to overcome compile error:
http://stackoverflow.com/questions/12772927/specifying-an-online-image-in-sphinx-restructuredtext-format