cht: add a:defRPr with c:rich
Prior `_new_rich()` method neglected to add the `a:RPr/a:defRPr` subtree
under the `a:p` element. It turns out that inheritance of font
characteristics is sometimes disrupted if the defRPr element is missing.
Change `_new_rich()` to add that, and update to parsing mechanism for
adding a subtree while we're in there.