fix: improve handling of media click-action
An imported audio (or video I suppose) media item produces a `<p:pic>`
element showing the "speaker" audio image. This shape has a click-action
which plays (or stops) the audio when clicked on in presentation-mode.

This `ppaction://media` click-action is not a "registered" click-action
and it has no counterpart in the `MsoPpAction` enumeration (PP_ACTION in
`python-pptx`). Previously this would cause `ActionSetting.action` to
raise `KeyError` when accessed for this type of shape.

Change the behavior so `PP_ACTION.NONE` is returned in this case
instead.