fix: Python 2.7 and 3.6 test errors
These resulted mostly from pulling `from __future__ ...` imports out
wholesale. A few were actually still needed.

Also, setting autospec=True by default on `method_mock()` encountered a
but that was present in 3.6 but not 2.7 or 3.8, so we need to set some
method-mocks to autospec=False even though that works properly now in
3.8.