Adding interface to wide.pptx template
Presentation now has an optional keyword template with valid values of default and wide. Other options can be added later. Instead of checking for known options (default and wide), it checks for an existing path. If the path does not exist, the error informs the user that default and wide were options.